+++ examples/tcp/common.rs
@@ -1,6 +1,0 @@
-use serde::{Serialize, Deserialize};
-
-#[derive(Serialize, Deserialize)]
-pub enum Message {
-    Greetings(String),
-}