+++ src/lib.rs
@@ -5,10 +5,10 @@ mod encoding;
 mod poll;
 mod driver;
 mod engine;
+mod adapters;
+mod remote_addr;
 
-pub mod remote_addr;
 pub mod adapter;
-pub mod adapters;
 pub mod network;
 pub mod events;