+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Release 0.7.1
 - Added WebSocket support based in `tungstenite-rs`
+- Added `Network::split()` function.
+- Join `udp` and `tcp` examples to make the `ping-pong` example with *WebSocket* support.
+- Improved `RemoteAddr` traits.
 
 ## Release 0.7.0
 - Internal improvements in order to use one thread for all adapters.